+++ src/styles/color.ts
@@ -3,7 +3,7 @@ export const color = {
   orange: "#FF763B",
   red: "#F44336",
 
-  maruDefault: "#1470FF",
+  maruDefault: "#257CFF",
   maruHigh: "#3081FF",
   maruRow: "#0C57CB",
   maruHoverd: "#186EEE",
